backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
ecdsa==0.19.1
fastapi==0.118.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
groq==0.32.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.3
openai==0.28.0
packaging==25.0
passlib==1.7.4
preshed==3.0.10
propcache==0.4.1
pyasn1==0.6.1
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
sympy==1.14.0
thinc==8.3.6
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
yarl==1.22.0
